Fixed client packet

+++ core/src/main/java/simulation/GameObjects/ObjectEncoder.java
@@ -4,7 +4,7 @@ import io.netty.buffer.ByteBuf;
 
 public class ObjectEncoder {
     public static void encode(ByteBuf buf, GameObject object) {
-        buf.writeByte(object.getId());
+        buf.writeInt(object.getId());
         buf.writeByte(object.getType());
         object.write(buf);
    }
